+++ Twitter_Frontend_API/__init__.py
@@ -1,4 +1,5 @@
+# -*- coding: utf-8 -*-
 from .main import Client
 from .main import API
 
-__version__ = "1.0.4"
+__version__ = "1.1.1"